Add support for SIMD instructions.
By default, the CPU rendering mode now uses a special SIMD implementation
of the mandelbrot algorithm if the CPU supports it. The use of SIMD
instructions can be disabled using the config or the --no-simd command line
flag. This commit also refactors some other code, but it would have been
hard to split that into its own commit.